Consolidate bounded JSON streams (#3681)
