+++ changelog.d/unreleased/3681.fixed.md
@@ -1,0 +1,18 @@
+---
+category: fixed
+issues:
+  - 3681
+affected:
+  - src/CodeIndex/Mcp/BoundedJsonUtf8Stream.cs
+  - src/CodeIndex/Mcp/McpServer.cs
+  - src/CodeIndex/Mcp/AuditLogSink.cs
+  - tests/CodeIndex.Tests/McpServerTests.cs
+---
+
+## English
+
+- **MCP JSON byte-limit streams are shared across response and audit-log serialization (#3681)** — bounded UTF-8 serialization now uses one helper with tests for unsupported stream operations and partial byte-limit capture.
+
+## 日本語
+
+- **MCP の JSON バイト制限 stream を response と audit-log serialization で共有しました (#3681)** — bounded UTF-8 serialization は 1 つの helper を使うようになり、未対応 stream 操作と byte limit 超過時の部分 capture をテストで固定しました。
